requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.11.7
aiosignal==1.3.1
async-timeout==5.0.1
attrs==24.2.0
frozenlist==1.5.0
idna==3.10
multidict==6.1.0
propcache==0.2.0
py-cord==2.6.1
python-dotenv==1.0.1
typing_extensions==4.12.2
yarl==1.18.0
